advancedREST/requirements.txt
Flask==1.1.2
google-cloud-datastore==1.11.0
google-auth==1.14.0
requests==2.22.0
json2html==1.3.0